refact(auth/forms login-changePass): change the login and change password forms
