+++ actions/auth.ts
@@ -2,7 +2,7 @@
 import { createSession, deleteSession } from '@/app/lib/session';
 import PostLogin from '@/utils/posts/postSignin';
 import PutUpdateUser from '@/utils/puts/putUpdateUser';
-import IChangePassForm from '@/utils/types/changePassFormInterface';
+import IChangePassForm from '@/utils/types/auth/changePassFormInterface';
 import ILoginForm from '@/utils/types/loginFormInterface';
 
 export const HandleLogin = async (credentials: ILoginForm) => {
