+++ app/components/loginForm/index.tsx
@@ -1,64 +1,12 @@
 'use client';
 import { LoginFormArray } from '../../../utils/arraysforms/loginForm';
-import React, { useState } from 'react';
-import ILoginForm from '../../../utils/types/loginFormInterface';
-import PostLogin from '../../../utils/posts/postSignin';
-import { useUserContext } from '../context';
-import { useRouter } from 'next/navigation';
-import ILoginErrorForm from '../../../utils/types/loginFormErrorInterface';
-import loginValidation from '../../../utils/formValidation/loginValidation';
-import redirectionByRole from '../../../utils/ redirects/redirectByRole';
-import Swal from 'sweetalert2';
-import { HandleLogin } from '@/actions/auth';
+import React from 'react';
+import ILoginForm from '../../../utils/types/auth/loginFormInterface';
+import useLoginForm from './useLoginForm';
 
 const LoginForm = () => {
-  const router = useRouter();
-  const { setUser, setToken } = useUserContext();
-  const [LoginForm, setLoginForm] = useState<ILoginForm>({
-    email: '',
-    password: '',
-  });
 
-  const [LoginFormError, setLoginFormError] = useState<ILoginErrorForm>({
-    email: '',
-    password: '',
-  });
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setLoginForm({ ...LoginForm, [name]: value });
-  };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const errors = loginValidation(LoginForm);
-    setLoginFormError(errors);
-    if (errors.email || errors.password) return;
-
-    const data = await HandleLogin(LoginForm);
-
-    //const data = await PostLogin(LoginForm);
-    if (!data.error) {
-      setUser(data.user);
-      setToken(data.token);
-      await Swal.fire({
-        icon: 'success',
-        title: 'Bienvenido',
-        showConfirmButton: false,
-        width: '450px',
-        timer: 2000,
-      });
-      if (data.user) {
-        const redirection = redirectionByRole(data.user.role);
-        router.push(redirection);
-      }
-    } else {
-      Swal.fire({
-        icon: 'error',
-        title: 'Error',
-        text: 'Usuario o contraseña incorrectos',
-      });
-    }
-  };
+  const { LoginForm, LoginFormError, handleChange, handleSubmit } = useLoginForm();
 
   return (
     <>
